Console - Fix user deletion

+++ console/src/main/java/org/georchestra/console/ws/backoffice/users/UsersController.java
@@ -438,8 +438,9 @@ public class UsersController {
 
 		this.accountDao.delete(uid, request.getHeader("sec-username"));
 
-		// Also delete delegation
-		this.delegationDao.delete(uid);
+		// Also delete delegation if exists
+		if(this.delegationDao.findOne(uid) != null)
+			this.delegationDao.delete(uid);
 
 		ResponseUtil.writeSuccess(response);
 	}
